Move student/teacher selection away from edge of page and above content
